adjust Terraform configs to fix warning on the newer versions
Terraform 0.14 produces the warning

```
Terraform 0.13 and earlier allowed provider version constraints inside the
provider configuration block, but that is now deprecated and will be removed
in a future version of Terraform. To silence this warning, move the provider
version constraint into the required_providers block.
```

the change should work on Terraform 0.12 as well