Second attempt to fix issue with phantom blocks in DLT pipeline
Current problem reported in #1401 is caused by the behaviour of the complex data types
inside the blocks that use the `slice_set` as a type - when we make changes the phantom
blocks are generated in addition to required changes.  Adding a custom hash function
didn't help.

This PR attempts to fix that issue other way - by changing the type of `cluster` blocks
from the `TypeSet` to the `TypeList`.  This approach, besides fixing the original issue
will also help with decreasing the amount of plan changes when making changes in the
`cluster` block.

The side effect of this change is that when upgrading to a new version, Terraform will
show changes in the DLT pipeline resource even if no actual changes were made in the code -
this is a result of changing the order of the `cluster` blocks