Custom suppress diff for `storage` in `databricks_pipeline`
When the `storage` option isn't provided, DLT generates a temp one in the form of
`dbfs:/pipelines/<uuid>`.  But in the sequential apply the drift was detected, and
pipeline was recreated.  Marking it with `computed` or `suppress_diff` doesn't help
because it prevents the pipeline from the detection of change of `storage` from value
to null.