Clarify that service principals are possible (#1775)
In the rest of Databricks documentation, service principals are viewed as distinct entities from users. For example in the [databricks_permission resource](https://registry.terraform.io/providers/databricks/databricks/latest/docs/resources/permissions#access-control-argument). It is best to clarify that this possible here and to clarify how to reference it.