Exporter: fix issue with exporting of resources with simple name (#1891)
* Exporter: fix issue with exporting of resources with simple name

The `Add` function is called only from the `Emit` so it's safe to remove the check if
resource was already imported or not.

this fixes #1838

* Make unique names for secret scopes & secrets

It's possible to use mixed characters in names that will generate non-unique names of
secret scopes & secrets.  Like, for unique scopes `Test-scope` and `test_scope` will
generate the same name, so terraform will complain about it.  The solution - add MD5
checksum over the scope & secret names

* Restore Has check in the Add, but do it differently