Don't use `computed` for `pause_status` in `databricks_job` resource (#2696)
Right now, the `pause_status` in `schedule`, `continuous` and `trigger` blocks is marked
as `computed` because it could be omitted, and backend will provide a default value.  This
leads to following problems:

- it's not changed to default `UNPAUSED` if you remove it from TF code (same behaviour
  would be with `suppress_diff`) - so you can't unpause the job after you paused it until
  you explicitly set `pause_status` to `UNPAUSED`
- this field isn't exported by TF exporter (issue #2695)

The solution for these problems is to set default in the provider's source code.

Also added validators for `pause_status` in `continuous` & `trigger` blocks.

This fixes #2695