Suppress diff for `user_name` in `databricks_user` when the changes only in character case
Identity management team rolled out the changes for AWS workspaces to force user names be
lower-cased by default (to match behavior on GCP & Azure).  But these changes are causing
the permanent configuration drift because `user_name` is marked as `force_new`.  To
mitigate this problem, added a suppress diff function that ignores changes in case.