Don't rely on having `@` to check if it's user or SP (#2765)
* Fix: don't rely on having `@` to check if it's user or SP

There are some installations where user name isn't equal to email - for them our detection
of username vs SP ID is broken, and lead to some problems, like, generation of erroneous
`run_as` blocks in job definitions, etc.

This PR fixes this by validating a given string as UUID, and if it's matches, then uses it
as SP ID, and the rest as user name.  Changes affect following:

* `databricks_job` API - besides validation, it also doesn't fill the `RunAs` structure on
  read when creator user name == run as user name.
* `databricks_current_user` data source - should fix generation of `acl_principal_id`

* Update jobs/data_job_test.go

Co-authored-by: Miles Yucht <miles@databricks.com>

* Add integration test

---------

Co-authored-by: Miles Yucht <miles@databricks.com>