Exporter: fix emitting of group membership (#2868)
There was an inconsistency in emitting of the `databricks_group_member` - depending on
where it was emitted - from group, or from user/service principal, the generated name was
either with user/SP ID, or without.  This behavior was always there, but it wasn't visible
until we started to do a parallel processing of resources.

This PR fixes this behavior, and group membership is consistent now.