Force sending the `content` field when creating empty `databricks_workspace_file` resources
The issue was reported by customer who used the exporter to migrate the Azure workspace
that they exported many empty `__init__.py` files, but couldn't import them because
Terraform gave them an error.  The reason for this error is that in API spec the `content`
field is marked as `optional` and correspondingly the `omitempty` was generated in Go SDK,
but in fact this field is required by the API.

To work around this problem until API spec is fixed, the `Content` field was explicitly
added to the `ForceSendFields` array to send the empty content.