Remove diff suppression for job `run_as` block
Diff suppression meant that once a nested key was set (either `user_name` or
`service_principal_name`), it was unable to unset it.  If both are specified,
they are ignored.

This meant that once a job was configured with a `run_as` user, it was not
possible to configure it with a service principal, and vice versa.