Allow to override built-in `databricks_cluster_policy` resources (#2941)
* Allow to override built-in `databricks_cluster_policy` resources

Databricks is shipped with a number of built-in cluster policies that are inherited from
the cluster policy families.  For built-in cluster policies we need to provide an
override, not the `definition`.  We also need to handle deletes for them differently -
just remove the overrides because we can't remove built-in policies.  We also need to
distinguish cases when we override built-in policy from cases when we create a new policy
based on the cluster policy family.

This PR makes following changes:

* Discovers a list of cluster policies families dynamically instead of relying on
  hardcoded list of policies
* When creating a new `databricks_cluster_policy`, check if it's built-in policy or not,
  and if it's built-in, then update it with policy override instead of trying to create.
* When deleting a created `databricks_cluster_policy`, check if it's built-in policy or
  not, and if it's built-in, then reset policy overrides to empty string to return back to
  the original state.

This fixes #2935

* Add integration test for policy overrides