Exporter: Fix exporting of `parent` attribute for SQL objects
There is an edge case when SQL object (dashboard/query/alert) is in the user home folder,
not a subfolder - in this case instead of reference to a resource, we generated a
reference to the data source, and because `parent` is marked as `ForceNew` it lead to SQL
object recreation when importing existing SQL objects into the TF state.

The workaround is to check if the given parent folder is top-level user folder, and remove
the `parent` because by default SQL object is put into home folder of the user.