Exporter: rework handling of listings and interactive prompting
When configuring services to list, the old code was relying on the substring match, and
this lead to problems when we have overlapping matches, like, `uc` service matched
`uc-allowlist`, etc.  This PR changed to use a map to store services to list so we don't
have a problem with overlapping services.

Changes also were made to the interactive prompting - instead of asking for each resource
type separately, we now ask about service and show which resources belong to that service.