Fix Read operation in `databricks_catalog_workspace_binding` so we can do import
During work on exporter we found that Read operation for this resource doesn't work when
it's used from scratch, so we were need to explicitly fill the TF state with data.   This
also doesn't allow to import this resource.  This PR fixes the problem and uses data from
`Id` to retrieve all necessary information.