Exporter: Don't generate resources with empty names (#3345)
Very often we emit resources that don't exist in backend, and this creates incorrect
resources without actual data. Typically this happens when objects are deleted, but still
referenced from jobs, dashboards, etc.

This PR skips generation of resources where the `name` is required attribute but it's
empty in the state, meaning that resource doesn't exist on the backend.  This is done for
the following resources:

* `databricks_sql_query`
* `databricks_sql_endpoint`
* `databricks_sql_alert
* `databricks_catalog`
* `databricks_schema`
* `databricks_volume`
* `databricks_sql_table`
* `databricks_registered_model`