Delete Vector Search Endpoint when we get an error while waiting for creation
If we get an error while waiting for the endpoint creation we simply return that
error.  But when we got timeout, we didn't cleanup the endpoint and this lead that
resource isn't destroyed in the tests as backend continued to provision it.

This PR deletes vector search endpoint when we got creation error during the wait phase.
This behaviour is similar to the current behavior of clusters.