Exporter: modify grants for some UC objects to allow handling of other owners
If UC objects, such as, metastore, catalogs, schemas, volumes, etc. have an owner other
than current user, then it's impossible to create child objects during apply.  To mitigate
this, the list of grants is expanded/changed to allow creation of child objects by the
identity who does the export (it's done only on workspace level, and the same identity
must be used when doing import).