Exporter: modify grants for some UC objects to allow handling of other owners (#3357)
* Exporter: modify grants for some UC objects to allow handling of other owners

If UC objects, such as, metastore, catalogs, schemas, volumes, etc. have an owner other
than current user, then it's impossible to create child objects during apply.  To mitigate
this, the list of grants is expanded/changed to allow creation of child objects by the
identity who does the export (it's done only on workspace level, and the same identity
must be used when doing import).

* Add possibility to pass extra data together with resource

Primary use - handling of ownership in the grants

* Improve handling of other owners

* check if the current user is already owner of an object, or no owner is specified
* set minimum permissions to make it working - like, `CREATE_SCHEMA` on catalogs, etc.