Exporter: don't generate references to ignored resources
Before this PR, exporter generated references to resources that are ignored (deleted DLT
pipelines in jobs, for example) - this happens because such object is still in the state,
so we can find it by ID, and generate a reference to it.  This PR adds a check if resource
is ignored or not, and if it's ignored, then don't generate a reference.