Exporter: generate resource data from the object status obtained during the listing operation
Before this PR, when we list workspace objects we just emitted the corresponding object
with just ID (path) -> this led to the need to call `Read` operation on the resource that
lead to additional `get-status` REST API call.  But `get-status` returns the same data
that are returned by `list` operation.  With this PR we generate resource data from object
status data returned by the `list` - this decreases the number of REST API calls necessary
to import a notebook (this is critical for handling of workspaces with huge number of
notebooks and workspace files).