Allow updating of comment in delta share resource
Why:
Based on the feedback in
https://github.com/databricks/terraform-provider-databricks/pull/3529#discussion_r1587552273,
that pull request needed to be extended to handle updates to `comment`
in the share resource

How:
Bit of a refactor to create `UpdateShareMeta` which takes in the ctx,
the resource information, client and the two different values in the
share. If only th `owner` is changed, then a call to set the owner is
set to the API. If the `owner` and `comment` are changed then both are
set to the API.

And likewise, handling just the `comment` change.

A test was added to demostrate that