Change Objects to a terraform Set
Why:
Currently, terraform treats the `object` blocks in a share as a
list. This results in the order of the `objects` mattering.

How:
This changes `Objects` to
`Objects   []SharedDataObject `json:"objects,omitempty"
tf:"slice_set,alias:object"`

`slice_set` sets the terraform schema type to Set and uses the default
hash function of of `HashSchema`

```
if setFunc == nil {
			// Default set function uses the schema to hash the whole value
			elem := s.Elem
			switch t := elem.(type) {
			case *Schema:
				setFunc = HashSchema(t)
			case *Resource:
				setFunc = HashResource(t)
			default:
				panic("invalid set element type")
			}
		}
		return &Set{F: setFunc}
```

`HashSchema` function calls `SerializeValueForHash` which looks at all
the fields and composes a hash for each of the types in the
`object`. This will capture all changes.

To keep the behavior of removing existing shares that *are not* in the
terraform state, this change also looks at existing shares from the
api call and compares them to the shares in the desired terraform
state. If the share is not present in the terraform state, then it is
removed.