[Exporter] Improve code generation for SQL Endpoints (#3764)
## Changes
<!-- Summary of your changes that are easy to understand -->

Changes include:

- Always generate `enable_serverless_compute` attribute for
`databricks_sql_endpoint` and `databricks_sql_global_config` - this is
required when there is a mix of serverless & non-serverless SQL
endpoints
- Don't generate `channel` when it's empty or is current
- Don't generate `tags` block when there are no tags
- Importable can define `ShouldGenerateField` function to specify if the
code for a field should be generated even if it's `omitempty` and having
the zero or default value. This was required for generation of
`enable_serverless_compute = false` that is marked as `omitempty`

Fixes #3758

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK