[Exporter] Adding more retries for SCIM API calls
There were issue when SCIM API failed to return list of users (because of the incorrect
status code or something like that, so Go SDK didn't retry it) and this lead to marking
all workspace objects (notebooks/files/...) as ignored because user wasn't found.  Also,
when the list of users wasn't retrieved, the process didn't stop, and we generated
incomplete Terraform code.

This PR improves the situation by following:

- Adding retries around Users/SPs/Groups SCIM API calls
- Exit with the error code when we aren't able to fetch the list of the users or service principals