[Dependency] Bump databricks-sdk-go to 0.44.0 (#3896)
## Changes
This PR bumps the Databricks SDK for Go to v0.44.0.

This version removes the Impl()/WithImpl() methods of the SDK. The
cachedMe implementation is slightly dependent on this interface, so I
changed how it is initialized to work with the new version of the SDK.
Additionally, there were two tests depending on this and one line in the
exporter. The tests are probably left-over from when the provider
contained its own client implementation, but now that it uses the SDK's
client for these resources, they can be removed.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK