[Internal] Fix integration tests (#3902)
## Changes
TestAccSQLEndpoint and TestAccWorkspaceConfFullLifecycle are currently
failing on `main`. The former fails in GCP because @ is not allowed in a
tag, so I've replaced it with _at_. The latter fails because of some
checks of capitalization which I've fixed in this PR.

Finally, TestMwsAccNetworkConnectivityConfig is failing on main but is a
regression server-side. The test passed on commit
7a2fc352f77101b89c3692ab34c82606605cceed when it ran on August 8, but it
now fails on the same commit.

## Tests
Ran these integration tests locally, they pass.

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK