[Internal] Upgrade `staticcheck` to v0.5.1 to get Go 1.23 support (#3931)
## Changes
<!-- Summary of your changes that are easy to understand -->

The old version (v0.4.0) is crashing when used with Go 1.23, so I was
need to bump the version and fix new findings (use of `fmt.Errorf`
instead of direct creation of errors from string).


## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK