[Fix] Force send fields for settings resources (#3978)
## Changes
<!-- Summary of your changes that are easy to understand -->

There was a problem that `databricks_automatic_cluster_update_setting`
wasn't applied because `minutes` attribute was set to 0, and it wasn't
sent in the JSON payload because it's marked as `omitempty` in the API.
After analysis, it was found that there are other places where we need
to send data but the corresponding fields were marked as `omitempty`.

This PR includes following changes:

- in 3 resources force sending of boolean and numeric fields with 0
values
- add customize schema for
`databricksenhanced_security_monitoring_setting`,
`databricks_automatic_cluster_update_setting` and
`databricks_enhanced_security_monitoring_setting` as fields are declared
as required in documentation/API, but not enforced in Terraform because
they are marked as `omitempty`
- expand documentation to include necessary blocks

Resolves #3976

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [x] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [x] using Go SDK