[Fix] Fix `TestAccClusterResource_WorkloadType` (#3989)
## Changes
`TestAccClusterResource_WorkloadType` merged with a small bug that
caused it to fail deterministically, asserting that the workload_type
had a clients block in state even when it wasn't specified in the
config, which is not expected. This PR changes the assertion on state to
verify that the block is not present in state when removed from the
config.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK