[Fix] Refactor `databricks_permissions` and allow the current user to set their own permissions (#3956)
## Changes
In
https://github.com/databricks/terraform-provider-databricks/commit/c441517af5ed3f2c6d793d64d5cf5d4e1ca0dc68,
a check was added to prevent users from assigning any permissions for
themselves in `databricks_permissions`. This unfortunately makes it
impossible for users to legitimately assign themselves as the owner of
certain resources, such as jobs, if they are currently owned by a
different principal. This PR removes this unnecessary restriction. If
the user requests to set permissions for an object in a way that is
incompatible with the object, such as removing themselves as owner, the
failure will be propagated from the backend to the user instead. This
does not make any changes to the way the ownership ACLs are set up (e.g.
for resources that require an owner, like jobs, if the Terraform user
did not specify an owner, the provider will still set the current user
as the owner).

This PR also refactors the permissions resource substantially. The logic
for implementing each resource type's permissions, including the field
name, object type and resource-specific modifications, are all colocated
with the resource's own definition. The type encapsulating this is
called`resourcePermissions`. As a result, the control flow is easier to
follow:
* Read reads from the API, customizes the response in a
resource-specific way, maps the response to the TF state representation,
and stores, or marks as deleted if there are no permissions.
* Create and Update read the desired permissions from ResourceData,
perform some validation, perform resource-specific, then puts the update
with an owner if not specified.
* Delete resets the ACLs to only admins + resource-specific
customizations.

Customizations are defined in the permissions/read and
permissions/update packages. For update, a mini expression language is
defined to support conditional application of customizations.

Lastly, this PR also migrates the resource to the Databricks SDK.

Fixes #2407.

## Tests
This PR adds integration test coverage for the `databricks_permissions`
resource for nearly all supported resource types. I wasn't able to run
the integration test for `authorization = "passwords"` because
password-based auth is deprecated, nor for serving endpoints because of
a likely race condition.

Integration tests cover all permission levels, and all principal types.
Included is special edge case testing for root directory and all
registered models.

- [ ] `make test` run locally
- [x] relevant change in `docs/` folder
- [x] covered with integration tests in `internal/acceptance`
- [x] relevant acceptance tests are passing
- [ ] using Go SDK