[Fix] Fix Permissions Dashboard Test (#4071)
## Changes
<!-- Summary of your changes that are easy to understand -->
It seems like we need to specify the `serialized_dashboard` field to be
able to create the dashboard. I have raised this concern with the
responsible team. But for unblocking our integration tests we can
specify this for now.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->
Integration Tests Passing
- [x] `make test` run locally
- [x] covered with integration tests in `internal/acceptance`
- [x] relevant acceptance tests are passing