[Fix] Set ID for online table resource if creation succeeds but it isn't available yet (#4072)
## Changes
<!-- Summary of your changes that are easy to understand -->
We should set the id right after creation and before waiting for online
table to be available. This is because in case when online table isn't
available, we still should have that resource in the state.

Also the timeout has been increased to 2x (I am going to following up
with online tables team for suitable timeout but since we have to do a
release, going ahead with small time increase should be good)

Note: We should add setting id right after creation for similar
resources to CONTRIBUTING guide (which I will do in a separate PR)

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->
Added unit test to check that pathway, id is set (which wasn't the case
before)
- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] relevant acceptance tests are passing
- [ ] using Go SDK

---------

Co-authored-by: Miles Yucht <miles@databricks.com>