[Internal] Update plugin framework schema to use ListNestedBlocks (#4040)
## Changes
<!-- Summary of your changes that are easy to understand -->
- To make sure we can release the resources with plugin framework that's
compatible with our sdkv2 resources, we need to use `ListNestedBlocks`
for struct types so that the `x.0.y` access pattern can be preserved,
also in the HCL we don't need to add extra `=` for blocks.
- To achieve this, we need corresponding changes in AttributeBuilder and
CustomizableSchema
- Updated tfsdk struct generations to generate list types for previous
struct types, so that the set and get functions will still work
- Updated the converter functions to cover cases that are structs in
`gosdk` but slices in `tfsdk`
- Updated acceptance tests to remove `=` for block types



## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [x] `make test` run locally
- [x] relevant change in `docs/` folder
- [x] covered with integration tests in `internal/acceptance`
- [x] relevant acceptance tests are passing
- [x] using Go SDK